HF 30 - Private groups

+++ src/components/modules/groups/MyGroups.jsx
@@ -5,6 +5,7 @@ import { Map } from 'immutable'
 import { api, formatter } from 'golos-lib-js'
 import tt from 'counterpart'
 
+import DialogManager from 'app/components/elements/common/DialogManager'
 import g from 'app/redux/GlobalReducer'
 import transaction from 'app/redux/TransactionReducer'
 import user from 'app/redux/UserReducer'
@@ -13,7 +14,7 @@ import DropdownMenu from 'app/components/elements/DropdownMenu'
 import Icon from 'app/components/elements/Icon'
 import LoadingIndicator from 'app/components/elements/LoadingIndicator'
 import { showLoginDialog } from 'app/components/dialogs/LoginDialog'
-import { getGroupLogo, getGroupMeta } from 'app/utils/groups'
+import { getGroupLogo, getGroupMeta, getRoleInGroup } from 'app/utils/groups'
 
 class MyGroups extends React.Component {
     constructor(props) {
@@ -69,14 +70,44 @@ class MyGroups extends React.Component {
         }))
     }
 
+    retireCancel = async (e, group, title, isPending) => {
+        e.preventDefault()
+        e.stopPropagation()
+        const { username } = this.props
+
+        let retireWarning
+        if (!isPending && group.privacy !== 'public_group') {
+            retireWarning = <span><br/>{tt('msgs_group_dropdown.joining_back_will_require_approval')}</span>
+        }
+        const res = await DialogManager.dangerConfirm(<div>
+            {(isPending ? tt('my_groups_jsx.are_you_sure_cancel') : tt('msgs_group_dropdown.are_you_sure_retire'))
+             + ' ' + title + '?'}
+            {retireWarning}</div>,
+            'GOLOS Messenger')
+        if (!res) return
+
+        this.props.groupMember({
+            requester: username, group: group.name,
+            member: username,
+            member_type: 'retired',
+            onSuccess: () => {
+                this.refetch()
+            },
+            onError: (err, errStr) => {
+                alert(errStr)
+            }
+        })
+    }
+
     showGroupSettings = (e, group) => {
         e.preventDefault()
         this.props.showGroupSettings({ group })
     }
 
-    showGroupMembers = (e, group) => {
+    showGroupMembers = (e, group, show_pendings) => {
         e.preventDefault()
-        this.props.showGroupMembers({ group })
+        const { name } = group
+        this.props.showGroupMembers({ group: name, show_pendings })
     }
 
     onGoGroup = (e) => {
@@ -85,7 +116,7 @@ class MyGroups extends React.Component {
     }
 
     _renderGroup = (group) => {
-        const { name, json_metadata } = group
+        const { name, json_metadata, pendings } = group
 
         const meta = getGroupMeta(json_metadata)
 
@@ -97,9 +128,22 @@ class MyGroups extends React.Component {
 
         const kebabItems = []
 
-        kebabItems.push({ link: '#', onClick: e => {
-            this.deleteGroup(e, group, titleShr)
-        }, value: tt('g.delete') })
+        const { username } = this.props
+        const { amOwner, amModer, amPending, amMember } = getRoleInGroup(group, username)        
+
+        if (amOwner) {
+            kebabItems.push({ link: '#', onClick: e => {
+                this.showGroupSettings(e, group)
+            }, value: tt('my_groups_jsx.edit') })
+            kebabItems.push({ link: '#', onClick: e => {
+                this.deleteGroup(e, group, titleShr)
+            }, value: tt('g.delete') })
+        }
+        if (amMember || (amModer && !amOwner)) {
+            kebabItems.push({ link: '#', onClick: e => {
+                this.retireCancel(e, group, titleShr, amPending)
+            }, value: tt('msgs_group_dropdown.retire') })
+        }
 
         return <tr key={name}>
             <Link to={'/' + name} onClick={this.onGoGroup}>
@@ -109,19 +153,33 @@ class MyGroups extends React.Component {
                 </td>
                 <td className='group-buttons' onClick={(e) => {
                     e.preventDefault()
+                    e.stopPropagation()
                 }}>
+                    {amPending ? <button className='button hollow alert' title={
+                        amPending ? tt('msgs_group_dropdown.pending') : null} onClick={e => {
+                        this.retireCancel(e, group, titleShr, amPending)
+                    }}>
+                        <Icon name='cross' size='0_95x' />
+                        <span className='btn-title'>{amPending ? tt('msgs_group_dropdown.cancel') : tt('msgs_group_dropdown.retire')}</span>
+                    </button> : null}
+                    {(amModer && pendings) ? <button className='button hollow' onClick={e => {
+                        this.showGroupMembers(e, group, true)
+                    }} title={tt('group_members_jsx.check_pending_hint')}>
+                        <Icon name='voters' size='0_95x' />
+                        <span className='btn-title'>{tt('group_members_jsx.check_pending') + ' (' + pendings + ')'}</span>
+                    </button> : null}
                     <button className='button' onClick={e => {
                         this.showGroupMembers(e, group)
                     }}>
                         <Icon name='voters' size='0_95x' />
                         <span className='btn-title'>{tt('my_groups_jsx.members')}</span>
                     </button>
-                    <button className='button hollow' onClick={e => {
+                    {/*amOwner ? <button className='button hollow' onClick={e => {
                         this.showGroupSettings(e, group)
                     }}>
                         <Icon name='pencil' size='0_95x' />
                         <span className='btn-title'>{tt('my_groups_jsx.edit')}</span>
-                    </button>
+                    </button> : null*/}
                     {kebabItems.length ? <DropdownMenu el='div' items={kebabItems}>
                         <Icon name='new/more' size='0_95x' />
                     </DropdownMenu> : null}
@@ -145,7 +203,7 @@ class MyGroups extends React.Component {
                     {tt('my_groups_jsx.empty')}
                     {tt('my_groups_jsx.empty2')}
                     <a href='#' onClick={this.createGroup}>
-                        {tt('my_groups_jsx.create')}
+                        {tt('my_groups_jsx.create')}.
                     </a>
                 </div>
             } else {
@@ -175,6 +233,7 @@ class MyGroups extends React.Component {
                </div>
                {button}
                {groups}
+               {hasGroups ? <div style={{ height: '50px' }}></div> : null}
         </div>
     }
 }
@@ -182,12 +241,12 @@ class MyGroups extends React.Component {
 export default connect(
     (state, ownProps) => {
         const currentUser = state.user.getIn(['current'])
-        const currentAccount = currentUser && state.global.getIn(['accounts', currentUser.get('username')])
+        const username = currentUser && currentUser.get('username')
         const my_groups = state.global.get('my_groups')
 
         return { ...ownProps,
             currentUser,
-            currentAccount,
+            username,
             my_groups,
         }
     },
@@ -203,8 +262,8 @@ export default connect(
         showGroupSettings({ group }) {
             dispatch(user.actions.showGroupSettings({ group }))
         },
-        showGroupMembers({ group }) {
-            dispatch(user.actions.showGroupMembers({ group }))
+        showGroupMembers({ group, show_pendings }) {
+            dispatch(user.actions.showGroupMembers({ group: ['my_groups', group], show_pendings }))
         },
         deleteGroup: ({ owner, name, password,
         onSuccess, onError }) => {
@@ -231,6 +290,35 @@ export default connect(
                     if (onError) onError(err, errStr)
                 },
             }));
-        }
+        },
+        groupMember: ({ requester, group, member, member_type,
+        onSuccess, onError }) => {
+            const opData = {
+                requester,
+                name: group,
+                member,
+                member_type,
+                json_metadata: '{}',
+                extensions: [],
+            }
+
+            const plugin = 'private_message'
+            const json = JSON.stringify(['private_group_member', opData])
+
+            dispatch(transaction.actions.broadcastOperation({
+                type: 'custom_json',
+                operation: {
+                    id: plugin,
+                    required_posting_auths: [requester],
+                    json,
+                },
+                username: requester,
+                successCallback: onSuccess,
+                errorCallback: (err, errStr) => {
+                    console.error(err)
+                    if (onError) onError(err, errStr)
+                },
+            }));
+        },
     })
 )(MyGroups)
